[MISC][WK] move calendar into shared folder and redefined type for standAlone & internal
